Process inlets outlets (#24)
Contains two notebooks with structures at RWS boundaries:
- `nl-kunstwerken.ipynb` to create a GeoPackage from the data in
https://samenwerkenaankunstwerken.nl/#. Writes to:
https://deltares.thegood.cloud/f/54241
- `uitlaten_inlaten.ipynb` to create a HyDAMO GeoPackage from our own
collection of inlets_outlets. Writes to:
https://deltares.thegood.cloud/f/52625

`environment.yml` is moved and considered to be a project-yml. For now
we only need openpyxl as extra so we can read Excel-files in Pandas.
Notebooks can run with this environment, after including hydamo from the
src/hydamo dir added by:

```
pip install -e .
`` 

In the activated enviroment