require Python 3.10 and associated ruff updates (#39)
This is a bit of a mixed bag of changes similar to which we merged in
https://github.com/Deltares/Ribasim last weeks.

First of all we now require at least Python 3.10. Before this was 3.9.
The differences are not too big, but I don't think we need to spend out
time supporting old Python versions at this moment, 3.10 is sufficiently
old.

Since we now tell ruff to target Python 3.10, and ask it to give syntax
upgrade suggestions, we fix those. Most of this is nicer typing
annotations, using e.g. `dict` instead of `from typing import Dict`. The
ruff updates also includes some configuration to make it work well with
notebooks.

Lastly I removed the typos check. I don't think it is suitable for this
repository as there are many Dutch terms used in the code and docs.