Support Ribasim checkout (#85)
Fixes #84.

This adds a Pixi `dev` environment, which points to the Ribasim
repository, which is expected to be cloned next to this repository. To
launch VSCode in this environment, use `open-vscode-dev.bat`. I added
`open-vscode.bat` as well to open the default environment, which points
to the latest ribasim release.


https://pixi.sh/latest/reference/configuration/#the-feature-and-environments-tables