Remove outdated parametrize notebooks which are not used atm
In this period, we focused on coupling the AGV model to the national network while ensuring the AGV model implements functionalities for water supply and drainage situations. Initially, we achieved this using a single notebook. Next, we will scale up and implement it across other waterschappen. In the near future, we plan to streamline this by calling a single Python file instead of ten similar notebooks, which should reduce maintenance time. Hence, the other other notebooks have been deleted.