Fix ruff errors on main (#121)
Fixes a part of #120. The goal for this PR is to resolve all 53 open
ruff errors on main, without adding exceptions. Currently the commits in
this PR bring it down to 29. The remaining errors are all of the form
`F821 Undefined name`, where the name is:

- dm_netwerk_path
- gdf_rhws
- model_name
- output_folder
- ribasim_toml
- stored_trc

For various files. Will look with @rbruijnshkv to resolve the rest.

---------

Co-authored-by: rbruijnshkv <r.bruijns@hkv.nl>