Disconnect ribasim_lumping for now
Fixes #122.

At this point we'll work with the already uploaded results of ribasim_lumping and patch those up for use with the current Ribasim version, like for instance done in #129.

That means for now we don't need to update this code to run. I don't delete the code because we probably want to use it later on and may need to look at the implementation. This does mean we cannot `import ribasim_lumping` within the pixi environment.