Make AGV preprocessing reproducible
Instead of local paths, this uses the CloudStorage class from the `ribasim_nl` package to load AGV data, such that this code can also run on other machines.

I uncommented the first processing part since that was needed for the second part to run.

This also adds an upload of the processed data to the cloud to `AmstelGooienVecht/verwerkt/preprocessed.gpkg`, so the result of this step is directly accesible as well.

It should not change and results, but it's good to go over this with @rbruijnshkv tomorrrow.