Deduplicate shortest path code (#176)
The code was identical except for comments and print statements. So now
we move it all into a single function that is part of the package, so
the individual scripts are just a few lines.

Only AGV was an outlier since it had a lot of code after the shared
code. Perhaps @rbruijnshkv can comment on if that is needed or not.
Another difference I noticed is this, AGV has this set to 5:

```
        distance_interval = 50  # The distance interval you want to segment the lines at
```

This only deduplicates, I didn't want to start modifying the code at the
same time. So I cannot run it since it relies on local paths that I
don't have.