Add pixi auto update action (#259)
This is the same as what we use in the Ribasim repo. It updates the our
pixi lock file once a month to make sure we are using recent packages.

I'm also changing the pre-commit update on the same time, such that
there is less chance of a difference in pre-commit results on CI and
locally.

@evetion can you make a CI_PR_PAT for this repo as well? Or would it
make sense to have such a PAT org-wide?