Take `Basin / time` out of the database (#371)
Currently all models put all tables in the database. This results in
databases of several hundred MB. For large tables it is much more
efficient to store them in a separate Apache Arrow file.

I noticed it took 5 minutes to initialize a model due to it reading a
large `Basin / time` table from the database. I could improve this in
https://github.com/Deltares/Ribasim/pull/2533 but with this PR I propose
to always force this table outside of the database, by hacking it into
`ribasim_nl.Model.__init__`. To avoid cluttering the top level with too
many files next to the TOML, I also set `input_dir = "input"` so the
database and Arrow input files are stored there.

One downside of this is that because
https://github.com/Deltares/Ribasim/issues/318 is still open, this means
we cannot directly see `Basin / time` in QGIS, but we can drag and drop
Arrow files into QGIS to inspect them, if they are not too large.