HWS edits (#376)
I wanted to run the HWS steps, but got stuck at `2_basins.py` with
`ValueError: line with index 53 does not start in a polygon`.
EDIT: this is fixed after the latest version of
`categorie_oppervlaktewater.gpkg` got uploaded.

This has a few non-functional changes, like removing `engine="pyogrio"`
since that is the default if it is installed (and it is).