Update to ribasim v2025.5.0 (#383)
For changes see
https://github.com/Deltares/Ribasim/releases/tag/v2025.5.0

This also includes the renaming of `greater_than` to `threshold_high` in
DiscreteControl, done in https://github.com/Deltares/Ribasim/pull/2564.

This could break some code, especially if we read data with columns like
`greater_than` that we directly put into DiscreteControl, such as some
of the data in `Rijkswaterstaat/verwerkt/kunstwerken`. In this case it
is probably easiest to just update the data.