Use DVC for tracking data and pipelines (#126)
See https://dvc.org/doc, and https://www.gitops.tech/

Shortly, this is like git-lfs, but on steroids. DVC can specify data
pipelines, registering the required input/output for each stage in a
storage platform (in our case dav). This should enable us to reproduce
every build we publish, by storing the hash of the required data next to
the code in git.

```mermaid
flowchart TD
   id["Input datasets"] -- generate --> id2[Ribasim models] -- test -->  id3[Test outputs]
	id2 -- stitch --> id4[Combined Model] -- test --> id5[Test output]
	id5 -- on git tag --> id6[Release model]
```

In the long term, DVC would replace the cloud.py codebase. This requires
full automatisation of our code, so we're probably not there yet.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>