DVC add Noorderzijlvest, disable samenvoegen (#451)
Most of the diff here is coming from adding this pre-commit hook, that
we also use in the Ribasim repo:
```yaml
  - repo: https://github.com/google/yamlfmt
    rev: v0.20.0
    hooks:
      - id: yamlfmt
```

I tried to fix the samenvoegen stage but ran into the problem that the
input models have inactive cells that cannot be migrated automatically.
If we generate new input models those will be fixed since I did
https://github.com/Deltares/Ribasim-NL/pull/450/changes/3acc028f9fcd71b5aa9de34c7a2041477e80673b.
For now I just added `frozen: true` to the samenvoegen stage, so it
won't run in CI. I also removed the GitHub Actions DVC CI since we have
it in TeamCity now (needed for longer runtimes and easier access to
ribasim binaries).

I ran `aa_en_maas` locally and did a DVC push, let's see if we can get
that one running.
Added `noorderzijlvest` as well.